+++ sn_cli/src/subcommands/wallet.rs
@@ -77,6 +77,9 @@ pub enum WalletCmds {
     },
     /// Receive a transfer created by the 'send' command.
     Receive {
+        /// Read the encrypted transfer from a file.
+        #[clap(long, default_value = "false")]
+        file: bool,
         /// Encrypted transfer.
         #[clap(name = "transfer")]
         transfer: String,
@@ -117,7 +120,6 @@ pub(crate) async fn wallet_cmds_without_client(cmds: &WalletCmds, root_dir: &Pat
             Ok(())
         }
         WalletCmds::Deposit { stdin, cash_note } => deposit(root_dir, *stdin, cash_note.clone()),
-        WalletCmds::GetFaucet { url } => get_faucet(root_dir, url.clone()).await,
         cmd => Err(eyre!("{cmd:?} requires us to be connected to the Network")),
     }
 }
@@ -130,7 +132,8 @@ pub(crate) async fn wallet_cmds(
 ) -> Result<()> {
     match cmds {
         WalletCmds::Send { amount, to } => send(amount, to, client, root_dir, verify_store).await?,
-        WalletCmds::Receive { transfer } => receive(transfer, client, root_dir).await?,
+        WalletCmds::Receive { file, transfer } => receive(transfer, file, client, root_dir).await?,
+        WalletCmds::GetFaucet { url } => get_faucet(root_dir, client, url.clone()).await?,
         WalletCmds::Pay {
             path,
             batch_size: _,
@@ -169,7 +172,7 @@ fn balance(root_dir: &Path) -> Result<NanoTokens> {
     Ok(balance)
 }
 
-async fn get_faucet(root_dir: &Path, url: String) -> Result<()> {
+async fn get_faucet(root_dir: &Path, client: &Client, url: String) -> Result<()> {
     let wallet = LocalWallet::load_from(root_dir)?;
     let address_hex = hex::encode(wallet.address().to_bytes());
     let url = if !url.contains("://") {
@@ -184,7 +187,7 @@ async fn get_faucet(root_dir: &Path, url: String) -> Result<()> {
     let is_ok = response.status().is_success();
     let body = response.text().await?;
     if is_ok {
-        deposit_from_cash_note_hex(root_dir, body)?;
+        receive(body, false, client, root_dir).await?;
         println!("Successfully got tokens from faucet.");
     } else {
         println!(
@@ -292,11 +295,18 @@ async fn send(
     Ok(())
 }
 
-async fn receive(transfer: String, client: &Client, root_dir: &Path) -> Result<()> {
+async fn receive(transfer: String, is_file: bool, client: &Client, root_dir: &Path) -> Result<()> {
+    let transfer = if is_file {
+        std::fs::read_to_string(transfer)?.trim().to_string()
+    } else {
+        transfer
+    };
+
     let transfer = match Transfer::from_hex(&transfer) {
         Ok(transfer) => transfer,
         Err(err) => {
             println!("Failed to parse transfer: {err:?}");
+            println!("Transfer: \"{transfer}\"");
             return Ok(());
         }
     };